pyproject.toml
[project]
name = "prototype-adiiva"
version = "0.1.0"
description = "Real-time voice AI agent with ≤ 2s end-to-end latency"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "deepgram-sdk>=3.7.0",
    "google-genai>=1.0.0",
    "elevenlabs>=1.9.0",
    "aiohttp>=3.10.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
